# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Specialist APP APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/en/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/orderStates": {
            "get": {
                "tags": [
                    "OrderState"
                ],
                "summary": "Get a listing of the OrderStates.",
                "description": "Get all OrderStates",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/OrderState"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrderState"
                ],
                "summary": "Store a newly created OrderState in storage",
                "description": "Store OrderState",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "OrderState that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/OrderState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/OrderState"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/orderStates/{id}": {
            "get": {
                "tags": [
                    "OrderState"
                ],
                "summary": "Display the specified OrderState",
                "description": "Get OrderState",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderState",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/OrderState"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrderState"
                ],
                "summary": "Update the specified OrderState in storage",
                "description": "Update OrderState",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderState",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "OrderState that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/OrderState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/OrderState"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrderState"
                ],
                "summary": "Remove the specified OrderState from storage",
                "description": "Delete OrderState",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderState",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "OrderState": {
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "name",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token\n *      Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}